feat: refined installer, now it creates a new folder

+++ src/main/java/taflgames/Installer.java
@@ -1,9 +1,12 @@
 package taflgames;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Collections;
 
 import org.yaml.snakeyaml.Yaml;
@@ -14,7 +17,8 @@ import org.slf4j.LoggerFactory;
  * A class used to create a file for the leaderboard.
  */
 public final class Installer {
-    private static final String PATH = System.getProperty("user.home");
+    private static final String SEP = File.separator;
+    private static final String PATH = System.getProperty("user.home") + SEP + ".tafl-games";
     private static final String FILE_NAME = "leaderboard.yaml";
 
     private static final Logger LOGGER = LoggerFactory.getLogger(Installer.class);
@@ -29,23 +33,42 @@ public final class Installer {
      * @return true if the file exists.
      */
     public static boolean doesFileExist() {
-        return new File(PATH + File.separator + FILE_NAME).isFile();
+        return Files.exists(Paths.get(PATH + File.separator + FILE_NAME));
+    }
+
+    /**
+     * Tests if the expected directory exists.
+     * @return true if the directory exists.
+     */
+    public static boolean doesDirectoryExist() {
+        return Files.isDirectory(Paths.get(PATH));
     }
 
     /**
      * Attempts to create the file for the leaderboard.
-     * @return ad
+     * @return true if the file was successfully created.
      */
     public static boolean createFile() {
-        /* final File folder = new File(PATH);
-        boolean b = folder.mkdir();
-        System.out.println("Created folder? " + b); */
-        final File leaderboardFile = new File(PATH + File.separator + FILE_NAME);
-        try (FileWriter writer = new FileWriter(leaderboardFile, StandardCharsets.UTF_8)) {
-            final Yaml yaml = new Yaml();
-            yaml.dump(Collections.emptyMap(), writer);
-        } catch (IOException e) {
-            LOGGER.error("Error in installer while trying to create folder", e);
+        if (!doesDirectoryExist()) {
+            try {
+                final File leaderboardDirectory = new File(PATH);
+                if (!leaderboardDirectory.mkdir()) {
+                    throw new FileNotFoundException();
+                }
+                LOGGER.info("Leaderboard directory successfully created.");
+            } catch (FileNotFoundException e) {
+                LOGGER.error("Could not create directory for leaderboard file.", e);
+            }
+        }
+        if (!doesFileExist()) {
+            final File leaderboardFile = new File(PATH + File.separator + FILE_NAME);
+            try (FileWriter writer = new FileWriter(leaderboardFile, StandardCharsets.UTF_8)) {
+                final Yaml yaml = new Yaml();
+                yaml.dump(Collections.emptyMap(), writer);
+                LOGGER.info("File 'leaderboard.yaml' at " + leaderboardFile.getAbsolutePath() + " successfully created.");
+            } catch (IOException e) {
+                LOGGER.error("Error in installer while trying to create file", e);
+            }
         }
         return doesFileExist();
     }
